+++ src/platforms/stub/time_stub.h
@@ -12,8 +12,8 @@
 extern "C" {
     uint32_t millis(void);
     uint32_t micros(void);
-    void delay(int ms);
-    void delayMicroseconds(int us);
+    void delay(unsigned long ms);
+    void delayMicroseconds(unsigned long us);
     void yield(void);
 }
